main graph test- add node (existing/not, with/out edges)

+++ edges.h
@@ -6,13 +6,13 @@
 
 typedef struct Edge {
     int src;
-    float w;
+    int w;
     int dest;
     struct Edge *prev_e;
     struct Edge *next_e;
 } edge, *p_edge;
 
-edge *gen_edge(int src_, float w_, int dest_);
+edge *gen_edge(int src_, int w_, int dest_);
 //char *to_str(p_edge e);
 int print_e(edge *e);
 
@@ -32,7 +32,7 @@ p_edge search_e(p_e_list pe_list, edge *e);
 
 void push_edge(p_e_list pe_list, edge *e);
 
-void push_Edge(p_e_list pe_list, int src_, float w_, int dest_);
+void push_Edge(p_e_list pe_list, int src_, int w_, int dest_);
 
 int pop_e_root(p_e_list pe_list);
